# api-specs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Fruition
  description: Fruition PoC 백엔드 API
  title: Fruition API
  version: 0.0.1
security:
- bearerAuth: []
...
paths:
...
  /api/workspaces/{workspace_id}/assets/{asset_id}/content:
    get:
      description: 워크스페이스 멤버에게 관리 이미지 bytes를 반환합니다.
      operationId: getContent
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: asset_id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
                format: binary
          description: 이미지 반환
        "304":
          content:
            '*/*':
              schema:
                type: string
                format: binary
          description: 캐시된 이미지 사용
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: asset 또는 워크스페이스를 찾을 수 없음
      summary: 문서 이미지 조회
      tags:
      - Document Assets
  /api/workspaces/{workspace_id}/chat/sessions:
    get:
      description: 가장 최근 메시지 순으로 정렬해 반환합니다.
      operationId: list_1
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatSessionListResponse"
          description: 조회 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
      summary: 채팅 세션 목록 조회
      tags:
      - Chat Sessions
    post:
      description: 워크스페이스당 최대 10개까지 생성할 수 있습니다.
      operationId: create_1
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionCreateRequest"
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatSessionResponse"
          description: 생성 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 세션 개수 제한 초과
      summary: 채팅 세션 생성
      tags:
      - Chat Sessions
  /api/workspaces/{workspace_id}/chat/sessions/{session_id}:
    delete:
      description: 워크스페이스에서 지정한 채팅 세션과 해당 세션의 메시지 기록을 삭제합니다.
      operationId: delete_1
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 채팅 세션 ID
        example: session_abc12345
        in: path
        name: session_id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: 삭제 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 세션 또는 워크스페이스를 찾을 수 없음
      summary: 채팅 세션 삭제
      tags:
      - Chat Sessions
    patch:
      description: 지정한 채팅 세션의 제목을 변경합니다. 목록 정렬 기준인 last_message_at은 바뀌지 않습니다.
      operationId: rename_3
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 채팅 세션 ID
        example: session_abc12345
        in: path
        name: session_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionRenameRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatSessionResponse"
          description: 변경 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 제목이 비었거나 255자를 초과함
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 세션 또는 워크스페이스를 찾을 수 없음
      summary: 채팅 세션 제목 변경
      tags:
      - Chat Sessions
  /api/workspaces/{workspace_id}/chat/sessions/{session_id}/messages:
    get:
      description: 세션 내 채팅 메시지를 생성 순서대로 반환합니다.
      operationId: getMessages
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 채팅 세션 ID
        example: session_abc12345
        in: path
        name: session_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatMessagesResponse"
          description: 조회 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 세션 또는 워크스페이스를 찾을 수 없음
      summary: 채팅 메시지 기록 조회
      tags:
      - Chat Sessions
...
  /api/workspaces/{workspace_id}/chat/sessions/{session_id}/wiki:
    post:
      description: 세션(full) 또는 선택 문답(partial)을 Markdown 원문 문서로 저장합니다. Ingest는 문서에서 별도로 요청합니다.
      operationId: exportToWiki
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 채팅 세션 ID
        example: session_abc12345
        in: path
        name: session_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatWikiExportRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatWikiExportResponse"
          description: 채팅 원문 문서 저장 완료 또는 기존 문서 반환
      summary: 채팅 Wiki page화
      tags:
      - Chat Sessions
  /api/workspaces/{workspace_id}/chat/sessions/{session_id}/wiki/preview:
    post:
      description: 세션을 llmPipeline 입력용 Markdown으로 직렬화해 결과만 반환합니다. 저장/파이프라인 호출은 하지 않습니다.
      operationId: previewWikiMarkdown
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 채팅 세션 ID
        example: session_abc12345
        in: path
        name: session_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            text/plain;charset=UTF-8:
              schema:
                type: string
          description: OK
      summary: "[임시] 채팅 Wiki page화 Markdown 미리보기"
      tags:
      - Chat Sessions
...
  /api/workspaces/{workspace_id}/wiki-schema/preview:
    post:
      description: Schema 규칙을 저장하지 않고 적용해 예상 Wiki 구조를 반환합니다.
      operationId: preview
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WikiSchemaPreviewRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WikiSchemaPreviewResponse"
          description: 미리보기 생성 성공
        "400":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/JsonNode"
                - $ref: "#/components/schemas/ErrorResponse"
          description: 잘못된 Schema 또는 입력
        "404":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/JsonNode"
                - $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
        "422":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/JsonNode"
                - $ref: "#/components/schemas/ErrorResponse"
          description: Schema 요청 검증 실패
        "503":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: llmPipeline 사용 불가
      summary: Wiki Schema 미리보기
      tags:
      - Wiki Schema
...
components:
  schemas:
...
    ChatMessageReference:
      type: object
      description: 답변의 근거가 된 원문 참조. 값이 없는 필드는 키 자체가 빠진다.
      properties:
        id:
          type: integer
          format: int64
          description: 참조 ID
          example: 1
        rank:
          type: integer
          format: int32
          description: 근거 순위(0부터). 낮을수록 강한 근거다.
          example: 0
        reference_type:
          type: string
          description: 참조 종류
        source_block_ids:
          type: array
          description: 근거가 된 원문 block ID 목록
          items:
            type: string
        source_document_id:
          type: string
          description: 근거가 된 원본 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        source_refs:
          type: array
          description: 문서·block 쌍으로 표현한 근거 위치 목록
          items:
            $ref: "#/components/schemas/ChatSourceRef"
        text:
          type: string
          description: 근거 구절 원문
    ChatMessageRelatedPageResponse:
      type: object
      description: 답변과 관련된 Wiki 페이지
      properties:
        depth:
          type: integer
          format: int32
          description: 그래프 탐색에서의 거리. 0이면 직접 매칭이다.
          example: 1
        page_type:
          type: string
          description: 페이지 종류
          example: Concept
        rank:
          type: integer
          format: int32
          description: 관련도 순위(0부터)
          example: 0
        relevance_score:
          type: number
          format: double
          description: 질문과의 관련도 점수
          example: 0.87
        role:
          type: string
          description: 이 페이지가 답변에서 맡은 역할
        slug:
          type: string
          description: URL에 쓰는 식별자
          example: search-indexing
        title:
          type: string
          description: 페이지 제목
          example: 검색 인덱싱
        wiki_page_id:
          type: string
          description: Wiki 페이지 ID
    ChatMessageResponse:
      type: object
      description: 채팅 메시지 한 건. 질문(user)과 답변(assistant)이 pair_id로 묶인다.
      properties:
        action:
          type: string
          description: AI가 고른 갈래. 화면이 이 값으로 편집 미리보기와 일반 답변을 나눈다. 질의 메시지는 키가 빠진다.
          enum:
          - chat_answer
          - conversation_reply
          - markdown_edit
          - markdown_create
          - clarify
          - reject
          - folder_organize
          - workspace_workflow
          - skill_authoring
          - skill_draft_proposal
          example: markdown_edit
        content:
          type: string
          description: 메시지 본문
        created_at:
          type: string
          format: date-time
          description: 생성 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
        error_message:
          type: string
          description: 실패했을 때의 사유. 성공이면 키가 빠진다.
        id:
          type: string
          description: 메시지 ID
        model:
          type: string
          description: 이 답변을 생성한 모델명
          example: gpt-5-nano
        pair_id:
          type: string
          description: 질문·답변을 묶는 ID. Wiki export에서 선택 단위로 쓴다.
        partial_wiki_page_ids:
          type: array
          description: 이 문답을 내보내 만들어진 Wiki 페이지 ID 목록
          items:
            type: string
        provider:
          type: string
          description: 이 답변을 생성한 LLM provider. 실행 시점 값이 snapshot으로 남는다.
          example: openai
        references:
          type: array
          description: 답변의 근거가 된 원문 참조 목록
          items:
            $ref: "#/components/schemas/ChatMessageReference"
        related_pages:
          type: array
          description: 답변과 관련된 Wiki 페이지 목록
          items:
            $ref: "#/components/schemas/ChatMessageRelatedPageResponse"
        role:
          type: string
          description: 발화 주체
          enum:
          - user
          - assistant
          example: assistant
        run_id:
          type: string
          description: Agent turn이 만든 메시지의 run ID. 승인 상태와 미리보기 본문을 이 run에서 읽는다. 질의 메시지는 키가 빠진다.
          example: agent_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        status:
          type: string
          description: 처리 상태. 비동기 질의는 완료 전 pending 상태로 먼저 나타난다.
          example: completed
        web_search_enabled:
          type: boolean
          description: 이 질의에 웹 검색이 허용됐는지. 실행 당시 값이 snapshot으로 남는다.
          example: false
    ChatMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          description: 세션의 메시지 기록. 오래된 것부터 온다.
          items:
            $ref: "#/components/schemas/ChatMessageResponse"
    ChatSessionCreateRequest:
      type: object
      properties:
        title:
          type: string
          description: 세션 제목. 생략하면 서버가 기본 제목을 붙인다.
          example: 검색 인덱싱 질문
    ChatSessionListResponse:
      type: object
      properties:
        sessions:
          type: array
          description: 워크스페이스의 채팅 세션 목록
          items:
            $ref: "#/components/schemas/ChatSessionResponse"
    ChatSessionRenameRequest:
      type: object
      properties:
        title:
          type: string
          description: 새 세션 제목
          example: 검색 인덱싱 질문
          maxLength: 255
          minLength: 0
      required:
      - title
    ChatSessionResponse:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          description: 생성 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
        id:
          type: string
          description: 채팅 세션 ID
          example: session_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        last_message_at:
          type: string
          format: date-time
          description: 마지막 메시지 시각(ISO-8601 UTC). 목록 정렬에 쓴다.
          example: 2026-08-13T04:25:24.371948Z
        title:
          type: string
          description: 세션 제목
          example: 검색 인덱싱 질문
    ChatSourceRef:
      type: object
      description: 근거가 가리키는 문서·block 쌍. 한 근거가 여러 문서를 참조할 수 있어 쌍으로 표현한다.
      properties:
        source_block_id:
          type: string
          description: 문서 안에서 근거가 있는 block ID
        source_document_id:
          type: string
          description: 근거가 있는 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
    ChatWikiExportRequest:
      type: object
      description: 채팅 내용을 Markdown 원문 문서로 저장하는 요청
      properties:
        pair_ids:
          type: array
          description: 내보낼 문답(pair) ID 목록. 비어 있을 수 없다.
          items:
            type: string
    ChatWikiExportResponse:
      type: object
      description: 채팅 원문 문서 저장 결과. Ingest는 별도로 요청한다.
      properties:
        exportDocumentId:
          type: string
          description: 생성된 export 문서 ID. 같은 내용이 이미 있으면 기존 문서 ID다.
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        status:
          type: string
          description: "saved는 저장 완료, skipped는 같은 내용이 이미 있어 건너뜀을 뜻한다."
          enum:
          - saved
          - skipped
          example: saved
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 분기 기준으로 쓰는 값. message가 아니라 이 값으로 분기한다.
          example: INVALID_REQUEST
        details:
          type: array
          description: 검증 실패(400 INVALID_REQUEST)일 때만 존재한다. 값이 없으면 키 자체가 빠지므로 null 체크가 아니라 존재 여부로 확인한다.
          items:
            $ref: "#/components/schemas/FieldError"
        message:
          type: string
          description: 사용자에게 그대로 보여도 되는 한국어 문구
          example: 요청 형식이 올바르지 않습니다.
    ErrorResponse:
      type: object
      description: 컨트롤러까지 도달한 요청이 실패했을 때 모든 API가 공유하는 에러 응답 형식. 인증 필터가 막는 401은 HttpStatusEntryPoint가 상태 코드만 내보내므로 이 본문이 없다.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 에러 상세
...
    FieldError:
      type: object
      properties:
        field:
          type: string
          description: 문제가 된 요청 필드명
          example: email
        reason:
          type: string
          description: 그 필드의 위반 사유
          example: email은 필수입니다.
...
    JsonNode: {}
...
    WikiSchemaFragmentsResponse:
      type: object
      description: 스키마를 쓰이는 곳별로 쪼갠 조각. 각 조각이 해당 단계의 프롬프트에 들어간다.
      properties:
        concept_markdown:
          type: string
          description: Concept 페이지 작성 규칙
        edit_markdown:
          type: string
          description: 편집 단계에만 적용되는 규칙
        global_markdown:
          type: string
          description: 모든 단계에 공통으로 들어가는 규칙
        ingest_markdown:
          type: string
          description: 문서 ingest 단계에만 적용되는 규칙
        query_markdown:
          type: string
          description: 질의 단계에만 적용되는 규칙
        template_markdown:
          type: string
          description: 페이지 템플릿 정의
      required:
      - concept_markdown
      - edit_markdown
      - global_markdown
      - ingest_markdown
      - query_markdown
      - template_markdown
    WikiSchemaIssueResponse:
      type: object
      description: 스키마 해석 중 발견한 문제 하나
      properties:
        category:
          type: string
          description: 문제 분류
        reason:
          type: string
          description: 왜 문제인지에 대한 설명
        section:
          type: string
          description: 문제가 발견된 섹션 이름. 특정할 수 없으면 null이다.
        severity:
          type: string
          description: 심각도. blocked면 활성화할 수 없고 unclear는 경고다.
          enum:
          - blocked
          - unclear
          example: unclear
        text:
          type: string
          description: 문제가 된 원문 구절
      required:
      - category
      - reason
      - severity
      - text
    WikiSchemaPreviewRequest:
      type: object
      description: Wiki 스키마 미리보기 요청. 저장하지 않고 결과만 돌려준다.
      properties:
        rawMarkdown:
          type: string
          description: 스키마를 뽑아낼 원본 Markdown
          example: |-
            # 설계

            ## 구성요소
          minLength: 1
      required:
      - rawMarkdown
    WikiSchemaPreviewResponse:
      type: object
      description: 저장하지 않고 스키마 해석 결과만 보여주는 응답
      properties:
        fragments:
          $ref: "#/components/schemas/WikiSchemaFragmentsResponse"
          description: 용도별로 쪼갠 스키마 조각
        has_blocked_issues:
          type: boolean
          description: 활성화를 막는 blocked 문제가 있는지 여부
          example: false
        issues:
          type: array
          description: 해석 중 발견한 문제 목록
          items:
            $ref: "#/components/schemas/WikiSchemaIssueResponse"
        preview_markdown:
          type: string
          description: 해석 결과를 사람이 읽도록 정리한 Markdown
      required:
      - fragments
      - has_blocked_issues
      - issues
      - preview_markdown
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http